apps/backend/workers/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
bullmq==2.16.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
decorator==5.2.1
distro==1.9.0
etelemetry==0.3.1
executing==2.2.1
filelock==3.20.0
fitz==0.0.1.dev2
frozenlist==1.8.0
genai==2.1.0
google-auth==2.43.0
google-genai==1.55.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
looseversion==1.3.0
lxml==6.0.2
matplotlib-inline==0.2.1
msgpack==1.1.2
multidict==6.7.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.3.5
openai==0.27.10
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
prompt-toolkit==3.0.52
propcache==0.4.1
prov==2.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
puremagic==1.30
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydot==4.0.1
pygments==2.19.2
pymupdf==1.26.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyxnat==1.6.4
rdflib==7.5.0
redis==6.4.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
semver==3.0.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
